chore: remove ClientDataGenerators.java GatherDataEvent.Server is not used from my knowledge, making it redundant to separate them
