+++ src/main/java/com/dimensiondelvers/dimensiondelvers/datagen/DataGenerators.java
@@ -3,17 +3,29 @@ package com.dimensiondelvers.dimensiondelvers.datagen;
 import com.dimensiondelvers.dimensiondelvers.DimensionDelvers;
 import net.minecraft.data.loot.LootTableProvider;
 import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
+import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.data.event.GatherDataEvent;
 import java.util.List;
 import java.util.Set;
 
-@EventBusSubscriber(modid = DimensionDelvers.MODID, bus = EventBusSubscriber.Bus.MOD)
+@EventBusSubscriber(modid = DimensionDelvers.MODID, bus = EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
 public class DataGenerators {
 
     @SubscribeEvent
-    public static void gatherData(GatherDataEvent.Server event) {
+    public static void gatherData(GatherDataEvent.Client event) {
+        event.createProvider(ModLanguageProvider::new);
+        event.createProvider(ModModelProvider::new);
+        event.createProvider((output, lookupProvider) -> new LootTableProvider(
+                output, Set.of(), List.of(new LootTableProvider.SubProviderEntry(ModBlockLootTableProvider::new, LootContextParamSets.BLOCK)), lookupProvider
+        ));
 
+        event.createProvider(ModRecipeProvider.Runner::new);
+
+        ModBlockTagProvider modBlockTagProvider = event.createProvider(ModBlockTagProvider::new);
+
+        event.createProvider((output, lookupProvider)
+                -> new ModItemTagProvider(output, lookupProvider, modBlockTagProvider.contentsGetter()));
     }
 }
